Don't throw on ValueGeneratedOnAdd without default
NpgsqlMigrationsSqlGenerator checks for the combination of
ValueGeneratedOnDefault without a default to infer sequence columns (and
db-generated UUIDs). If this combination existed on non-int and non-UUID
columns, an exception was thrown. The EFCore behavior seems to be to
ignore this case.

(cherry picked from commit 87ed91f9cca1b371388e448ed3004c8113a0724a)