Fixed check whether database exists
To check whether a database exists in NpgsqlDatabaseCreator, we open
a connection to that database. However, if that connection is pooled then
we may get back a connection that was broken in a previous database drop.
This has been plaguing the tests for a very long time.

Made the existence check use an unpooled connection.

(cherry picked from commit 74080091cc4421a497be6e2d54e1d637ccfc3f4f)