Serial redo and HiLo support
Replaced the previous serial column metadata system with more or less the
SQL Server provider's system, plus added HiLo sequence support. We now
have an NpgsqlValueGenerationStrategy on the properties (with a fallback
on the model), which can be either SerialColumn or SequenceHiLo.

Also added lots of missing builder extension methods and tests.

Closes #97, closes #5, closes #111