Make EnsurePostgresExtension work on 1.1
Versions 1.0.x had an EnsurePostgresExtension() extension method on model
builder which added a custom migration object. Unfortunately this turned
out to be unsupported by EFCore, and 1.1 switched to managing PostgreSQL
extensions via AlterDatabase annotations instead (see #101).

This change makes sure that the EnsurePostgresExtension() (and the
obsolete CreatePostgresExtensions()) that were scaffolded in 1.0.x will
still work in 1.1.0.

Related to #101