Don't scaffold default index method annotation
When scaffolding, we have an Npgsql-specific IndexMethod annotation to
represent PostgreSQL's index methods. Previously, we would always
output this annotation on indices, even when the method was the
default (btree); NpgsqlAnnotationCodeGenerator would elide it as by-
convention.

Although this is cleaner, it caused issues whenever two indices where
defined on the same column:
https://github.com/aspnet/EntityFrameworkCore/issues/11846#

We now scaffold the annotation only for non-default index methods as a
workaround; the issue should now affect much less people.

Fixes #228