Refactors subquery handling for complex patterns
This refactors `NpgsqlSqlTranslatingExpressionVisitor` to use
`VisitSubQuery` to delegate to `VisitArraySubQuery` which itself
delegates to more specialized virtual methods based on the
subquery's result operator. The goal is to allow the subquery
visits to be called recursively to handle more complex patterns.

For example, translating the array operator `<@` involves
identifying a subquery nested inside of a subquery. Rather than
duplicating identification logic, we want to use the standard
visitor approach and send a nested subquery back up to `Visit`.

Another consideration is the ability to compose result operators
within an `Any` or `All` result. These result operators are almost
exclusively designed to have a nested condition or subquery.