Begins work on a composite fragment translator
The plan is to use an expression fragment translator to
offload some of the work currently done in
`NpgsqlSqlTranslatingExpressionVisitor` and make complex
pattern translations more approachable.

Specifically, this makes it tractable to handle `Array.Exists<T>`
and `List<T>.Exists` for PostgreSQL array columns. This commit
includes new tests in various stages:

- Passing:
  - List_Exists_equals_with_literal_constant
  - List_Exists_less_than_with_literal_constant

- Failing (simple):
  - List_Exists_less_than_with_literal_constant

- Failing (complicated):
  - Array_Exists_equals_with_literal_constant
  - Array_Exists_equals_with_column_list_element
  - Array_Exists_less_than_with_literal_constant

The simple failure is related to sign inversion when switching
from C# to SQL patterns (e.g. 'y < 1' becomes '1 = ANY (...)`.

The complicated failures are due to missing rewrite support
for `Array.Exists<T>` and arrays generally. There is a built-in
`ExistsToAnyRewritingExpressionVisitor` which handles lists,
but it will need to be extended to support the same rewrite
for arrays.

Solving the above will provide exist translations for basic
query shapes. But this support is fragile at best until the
`NpgsqlArrayExpressionFragmentTranslator` class has a more
thorough treatment for array identification.

Currently, we just look for a `QuerySourceReferenceExpression`
and assume its a `MainFromClause`. This could be handled better
to allow for literal arrays, and possibly other set returning
functions that PostgreSQL would consider valid in an array
expression.