Updates to PgFunctionExpression
Changed the private constructor to internal to make it available
for expression visitors. Specifically, the following commit needs
to visit `PgFunctionExpression` instances that may have been
constructed in an early visitor.

The constructor pass-through logic was a bit tangled, so each
public constructor now passes directly to the internal one.

This saves a few redundant null checks, but mostly it makes it
clear which arguments are set to an internal default.

The `ToString` and `GetHashCode` logic has been cleaned up,
as well as adding an `IEquatable<T>` constraint. The thinking here
is that since the methods were already there, we might as well let
callers know.

Something to consider for future use would be whether or not
this expression should "own" its own SQL representation.

Basically, the `ToString` logic duplicates a lot of the same code
as `NpgsqlQuerySqlGenerator` (just without the instance/argument
visits).