Bring back support for PostgreSQL domains
Previously, if a column had an explicitly-specified datatype but a
mapping couldn't be found for that datatype/ClrType combination, we
returned null. This made it impossible to specify PostgreSQL domain
(alias) types as these aren't known to the type mapper.

If an explicitly-specified datatype isn't known, fall back to looking up
the mapping via ClrType only, ignoring the datatype. As long as the ADO
layer knows how to read/write the ClrType from/to the given datatype,
everything will work.

This is how the SqlServer mapper is implemented as well.

Fixes #454