Generates SQL for enum cast expressions (caution)
__However__, this introduces some problems. Enums used in binary
comparison operations (maybe others?) are cast to their underlying
system type in the expression tree.

So the fix in this commit is insufficient. We need to address this
from the `VisitBinary` methods, where either a custom expression
needs to be returned, or the whole comparison needs to be handled.

Otherwise, obvious looking code can't be handled. For example:

```c#
src.Where(x => x.MyInt == (int) myEnumVariable)
```