Fix user-defined range mappings
We used to eagerly create mappings for user-defined ranges in the
constructor of NpgsqlTypeMappingSource. This was too early, and did not
have access to mappings coming from plugins, so creating a user-defined
range over NodaTime types failed.

We know create mappings for user-defined ranges late, when FindMapping()
is called.

Fixes #688