Fix system column filtering in migrations
If a system column was present when creating a table, we filtered it
out by creating a copy of CreateTableOperation. We didn't copy
across all fields, so comments didn't get created.

Switched to the less brutal practice of mutating the operation in
place to remove columns instead of creating a copy.

Fixes #834