Automatically set sequence value after seeding
When inserting explicit IDs (during seeding), PostgreSQL doesn't update
the sequence value for auto-generated columns; this results in unique
constraint violations when the user next attempts to insert a new
entity.

Detect all auto-generated columns of tables which were seeded in a
migration, and reset their values to the current maximum.

Fixes #367