NpgsqlRelationalConnection owns newly created connection. Resets datasource when you decide to control the ownership of a connection. Used once in the setup UseNpgsql when you pass a connection. So no practical affect aside from enforcing no datasource.
The owned flag does not impact NpgsqlRelationalConnection CloneWith either, which seems like a bug to me in the RelationalConnection. Since the connection is opened in RelationalConnection it takes it as it should close it when Close is called. The dispose ignores it. Would make more sense to also enforce the ownership in Close and not only in Dispose, but that is not related to this repo.

This would however make it more correct since the datasource is reset and the RelationalConnection knows it has ownership of the connection when the connection is directly injected from a newly created connection which is not tracked elsewhere.